MakanakApi/requirements.txt
certifi>=2018.8.24
chardet>=3.0.4
coverage>=4.5.1
coveralls>=1.5.1
Django==2.1.2
django-cors-middleware>=1.5.0
django-extensions>=1.7.1
django-filter>=2.0.0
djangorestframework>=3.8.2
djangorestframework-jwt>=1.11.0
djangorestframework_simplejwt>1
django_rest_passwordreset>=1.0.0
docopt>=0.6.2
python-dotenv>=0.10.3
factory-boy>=2.11.1
Faker>=0.9.2
gunicorn
idna>=2.7
Markdown>=3.0.1
mysqlclient>=1.4
PyJWT>=1.6.4
python-dateutil>=2.7.3
pytz>=2018.5
requests>=2.19.1
six>=1.10.0
text-unidecode>=1.2
urllib3>=1.23
coreapi>=2.3.1
django-rest-swagger>=2.2.0
prospector>=1.2.0
autopep8==1.4
